requirements.txt
facebook_sdk==3.1.0
tweepy==3.8.0
numpy==1.16.2